requirements.txt
# Framework Web
streamlit==1.32.0

# Database Connector
mysql-connector-python==8.2.0

# Data Manipulation & Display
pandas==2.2.2

# Testing Framework
pytest==8.3.2
pytest-mock==3.14.0